Update non-background and synchronous usage exmaples
