+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,6 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.ZephrSampleClientApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -30,6 +29,14 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".BackgroundActivity"
+            android:exported="false"
+            android:theme="@style/Theme.ZephrSampleClientApp" />
+        <activity
+            android:name=".EmbeddedSolverActivity"
+            android:exported="false"
+            android:theme="@style/Theme.ZephrSampleClientApp" />
     </application>
 
 </manifest>
